+++ boilerplates/boilerplate-react-ts/renderer/_default.page.client.tsx
@@ -8,7 +8,7 @@ import type { PageContextBuiltInClient } from "vite-plugin-ssr/types";
 hydrate();
 
 async function hydrate() {
-  // For Client-side Routing we should use `useClientRouter()` instead of `getPage()`.
+  // For Client Routing we should use `useClientRouter()` instead of `getPage()`.
   // See https://vite-plugin-ssr.com/useClientRouter
   const pageContext = await getPage<PageContextBuiltInClient & PageContext>();
   const { Page, pageProps } = pageContext;