+++ boilerplates/boilerplate-react/renderer/_default.page.client.jsx
@@ -6,7 +6,7 @@ import { PageWrapper } from "./PageWrapper";
 hydrate();
 
 async function hydrate() {
-  // For Client-side Routing we should use `useClientRouter()` instead of `getPage()`.
+  // For Client Routing we should use `useClientRouter()` instead of `getPage()`.
   // See https://vite-plugin-ssr.com/useClientRouter
   const pageContext = await getPage();
   const { Page, pageProps } = pageContext;