+++ boilerplates/boilerplate-vue-ts/renderer/_default.page.client.ts
@@ -6,7 +6,7 @@ import type { PageContextBuiltInClient } from 'vite-plugin-ssr/types'
 hydrate()
 
 async function hydrate() {
-  // For Client-side Routing we should use `useClientRouter()` instead of `getPage()`.
+  // For Client Routing we should use `useClientRouter()` instead of `getPage()`.
   // See https://vite-plugin-ssr.com/useClientRouter
   const pageContext = await getPage<PageContextBuiltInClient & PageContext>()
   const app = createApp(pageContext)