+++ boilerplates/boilerplate-vue/renderer/_default.page.client.js
@@ -4,7 +4,7 @@ import { createApp } from './app'
 hydrate()
 
 async function hydrate() {
-  // For Client-side Routing we should use `useClientRouter()` instead of `getPage()`.
+  // For Client Routing we should use `useClientRouter()` instead of `getPage()`.
   // See https://vite-plugin-ssr.com/useClientRouter
   const pageContext = await getPage()
   const app = createApp(pageContext)