+++ vite-plugin-ssr/node/cli/bin.ts
@@ -16,7 +16,7 @@ cli
     '--root <path>',
     '[string] root directory of your project (where `vite.config.js` and `dist/` live) (default: `process.cwd()`)'
   )
-  .option('--client-router', 'serialize `pageContext` to JSON files for Client-side Routing')
+  .option('--client-router', 'serialize `pageContext` to JSON files for Client Routing')
   .option('--base <path>', '[string] public base path (default: /)')
   .option(
     '--parallel <numberOfJobs>',